Moved AdvancedInsight enums to separate file
